perbaikan penulisan invoice menghilang,pembuatan tampilan utama agar bisa dilihat user belum login

+++ routes/web.php
@@ -15,23 +15,26 @@ Route::get('/password/reset', [App\Http\Controllers\AuthController::class, 'show
 Route::post('/password/update', action: [App\Http\Controllers\AuthController::class, 'updatePassword'])->name('password.update');
 
 
-Route::get('/toko/buku/index', [App\Http\Controllers\BukuController::class, 'index'])->name('buku.index');
-Route::get('/toko/buku/create', [App\Http\Controllers\BukuController::class, 'create'])->name('buku.create');
-Route::post('/toko/buku', [App\Http\Controllers\BukuController::class, 'store'])->name('buku.store');
-Route::get('/toko/buku/{id}/edit', [App\Http\Controllers\BukuController::class, 'edit'])->name('buku.edit');
-Route::put('/toko/buku/{id}', [App\Http\Controllers\BukuController::class, 'update'])->name('buku.update');
-Route::delete('/toko/buku/{id}', [App\Http\Controllers\BukuController::class, 'destroy'])->name('buku.destroy');
+Route::get('/toko/buku/index', [App\Http\Controllers\BukuController::class, 'index'])->name('buku.index')->middleware('admin');
+Route::get('/toko/buku/create', [App\Http\Controllers\BukuController::class, 'create'])->name('buku.create')->middleware('admin');
+Route::post('/toko/buku', [App\Http\Controllers\BukuController::class, 'store'])->name('buku.store')->middleware('admin');
+Route::get('/toko/buku/{id}/edit', [App\Http\Controllers\BukuController::class, 'edit'])->name('buku.edit')->middleware('admin');
+Route::put('/toko/buku/{id}', [App\Http\Controllers\BukuController::class, 'update'])->name('buku.update')->middleware('admin');
+Route::delete('/toko/buku/{id}', [App\Http\Controllers\BukuController::class, 'destroy'])->name('buku.destroy')->middleware('admin');
 
 
-Route::get('/toko/sale', [App\Http\Controllers\SaleController::class, 'index'])->name('sale.index');
-Route::get('/toko/sale/create', [App\Http\Controllers\SaleController::class, 'create'])->name('sale.create');
-Route::post('/toko/sale', [App\Http\Controllers\SaleController::class, 'store'])->name('sale.store');
-Route::get('/invoice/{sale_id}', [App\Http\Controllers\SaleController::class, 'invoice'])->name('sale.invoice');
+Route::get('/toko/sale', [App\Http\Controllers\SaleController::class, 'index'])->name('sale.index')->middleware('admin');
+Route::get('/toko/sale/create', [App\Http\Controllers\SaleController::class, 'create'])->name('sale.create')->middleware('customer');
+Route::post('/toko/sale', [App\Http\Controllers\SaleController::class, 'store'])->name('sale.store')->middleware('customer');
+Route::get('/invoice/{sale_id}', [App\Http\Controllers\SaleController::class, 'invoice'])->name('sale.invoice')->middleware('customer');
 
-Route::get('/toko/hakakses', [App\Http\Controllers\HakaksesController::class, 'index'])->name('hakakses.index')/* ->middleware('admin') */;
-Route::get('/toko/hakakses/edit/{id}', [App\Http\Controllers\HakaksesController::class, 'edit'])->name('hakakses.edit')/* ->middleware('admin') */;
-Route::put('/toko/hakakses/update/{id}', [App\Http\Controllers\HakaksesController::class, 'update'])->name('hakakses.update')/* ->middleware('admin') */;
-Route::delete('/toko/hakakses/delete/{id}', [App\Http\Controllers\HakaksesController::class, 'destroy'])->name('hakakses.delete')/* ->middleware('admin') */;
+Route::get('/toko/hakakses', [App\Http\Controllers\HakaksesController::class, 'index'])->name('hakakses.index')->middleware('admin');
+Route::get('/toko/hakakses/edit/{id}', [App\Http\Controllers\HakaksesController::class, 'edit'])->name('hakakses.edit')->middleware('admin');
+Route::put('/toko/hakakses/update/{id}', [App\Http\Controllers\HakaksesController::class, 'update'])->name('hakakses.update')->middleware('admin');
+Route::delete('/toko/hakakses/delete/{id}', [App\Http\Controllers\HakaksesController::class, 'destroy'])->name('hakakses.delete')->middleware('admin');
 
-Route::get('layout/toko/buku', [App\Http\Controllers\BukuController::class, 'index'])->name('layout.toko.buku.index');
-Route::get('/dashboard', [App\Http\Controllers\DashboardController::class, 'index'])->name('dashboard');
+Route::get('layout/toko/buku', [App\Http\Controllers\BukuController::class, 'index'])->name('layout.toko.buku.index')->middleware('admin');
+Route::get('/dashboard', [App\Http\Controllers\DashboardController::class, 'index'])->name('dashboard')->middleware('customer');
+Route::get('/', [App\Http\Controllers\WelcomeController::class, 'index'])->name('welcome');
+
+
